Renamed directories structure.
