--- src/main/java/greencity/entities/Comment.java
@@ -1,37 +1,0 @@
-package greencity.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
-
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Comment {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String text;
-
-    @ManyToOne
-    private User user;
-
-    @ManyToOne
-    private Place place;
-
-    @ManyToOne
-    private Comment parentComment;
-
-    @OneToMany(mappedBy = "parentComment")
-    private List<Comment> comments = new ArrayList<>();
-}
