--- src/main/java/greencity/entities/Location.java
@@ -1,29 +1,0 @@
-package greencity.entities;
-
-import javax.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Location {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private Double lat;
-
-    @Column(nullable = false)
-    private Double lng;
-
-    @Column(name = "address", nullable = false)
-    private String address;
-
-    @OneToOne private Place place;
-}
