--- src/main/java/greencity/entities/SpecificationValue.java
@@ -1,29 +1,0 @@
-package greencity.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class SpecificationValue {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, length = 100)
-    private String value;
-    
-    @ManyToOne
-    private Place place;
-    
-    @ManyToOne
-    private Specification specification;
-
-}
