--- src/main/java/greencity/entities/UserOwnSecurity.java
@@ -1,29 +1,0 @@
-package greencity.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-
-@Entity
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class UserOwnSecurity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false)
-    private String password;
-
-    private Boolean isCheckedEmail = false;
-
-    @OneToOne
-    private User user;
-
-}
