--- src/main/java/greencity/entities/enums/PlaceStatus.java
@@ -1,7 +1,0 @@
-package greencity.entities.enums;
-
-public enum PlaceStatus {
-    PROPOSED,
-    DECLINED,
-    APPROVED
-}
