--- src/main/java/greencity/entities/enums/ROLE.java
@@ -1,5 +1,0 @@
-package greencity.entities.enums;
-
-public enum ROLE {
-    USER_ROLE, ADMIN_ROLE, MODERATOR_ROLE
-}
