--- src/main/java/greencity/entities/enums/WeekDay.java
@@ -1,5 +1,0 @@
-package greencity.entities.enums;
-
-public enum WeekDay {
-    MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY
-}
