+++ src/main/java/greencity/entity/Comment.java
@@ -1,0 +1,37 @@
+package greencity.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Comment {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String text;
+
+    @ManyToOne
+    private User user;
+
+    @ManyToOne
+    private Place place;
+
+    @ManyToOne
+    private Comment parentComment;
+
+    @OneToMany(mappedBy = "parentComment")
+    private List<Comment> comments = new ArrayList<>();
+}
