+++ src/main/java/greencity/entity/Rate.java
@@ -1,0 +1,29 @@
+package greencity.entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Rate {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    
+    @Column(nullable = false)
+    private Byte rate;
+    
+    @ManyToOne
+    private User user;
+
+    @ManyToOne
+    private Place place;
+
+}
