+++ src/main/java/greencity/entity/enums/PlaceStatus.java
@@ -1,0 +1,7 @@
+package greencity.entity.enums;
+
+public enum PlaceStatus {
+    PROPOSED,
+    DECLINED,
+    APPROVED
+}
