+++ src/main/java/greencity/entity/enums/ROLE.java
@@ -1,0 +1,5 @@
+package greencity.entity.enums;
+
+public enum ROLE {
+    USER_ROLE, ADMIN_ROLE, MODERATOR_ROLE
+}
