+++ src/main/java/greencity/entity/enums/WeekDay.java
@@ -1,0 +1,5 @@
+package greencity.entity.enums;
+
+public enum WeekDay {
+    MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY
+}
