+++ src/main/java/greencity/exception/ExceptionResponse.java
@@ -1,0 +1,30 @@
+package greencity.exception;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import java.util.List;
+import java.util.Map;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+
+public class ExceptionResponse {
+    private String message;
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    private List<String> details;
+    @JsonIgnore
+    private String timeStamp;
+    @JsonIgnore
+    private String trace;
+    @JsonIgnore
+    private String path;
+
+    public ExceptionResponse(Map<String, Object> errorAttributes) {
+        this.setPath((String) errorAttributes.get("path"));
+        this.setMessage((String) errorAttributes.get("message"));
+        this.setTimeStamp(errorAttributes.get("timestamp").toString());
+        this.setTrace((String) errorAttributes.get("trace"));
+    }
+}
