+++ src/main/java/greencity/repository/UserOwnSecurityRepo.java
@@ -1,0 +1,6 @@
+package greencity.repository;
+
+import greencity.entity.UserOwnSecurity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserOwnSecurityRepo extends JpaRepository<UserOwnSecurity, Long> {}
