+++ src/main/java/greencity/repository/UserRepo.java
@@ -1,0 +1,11 @@
+package greencity.repository;
+
+import greencity.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+/**
+ * Provides an interface to manage {@link User} entity.
+ */
+@Repository
+public interface UserRepo extends JpaRepository<User, Long> {}
