+++ src/main/java/greencity/service/PlaceService.java
@@ -1,0 +1,18 @@
+package greencity.service;
+
+import greencity.entity.Place;
+import greencity.entity.enums.PlaceStatus;
+import java.util.List;
+
+/**
+ * Provides the interface to manage {@code Place} entity.
+ * */
+public interface PlaceService {
+    /**
+     * Finds all {@code Place} with status {@code PlaceStatus}.
+     *
+     * @param placeStatus a value of {@link PlaceStatus} enum.
+     * @return a list of {@code Place} with the given {@code placeStatus}
+     */
+    List<Place> getPlacesByStatus(PlaceStatus placeStatus);
+}
