--- src/main/java/greencity/utils/readme.txt
@@ -1,1 +1,0 @@
-I use this file because of git does not push empty packages.
